Update dependency versions Increase targetSdk and compileSdk to 36 Revert back to the dynamic approach of setting versionName and versionCode

+++ app/src/main/java/com/thestrikingarts/ui/parentlayouts/ListScreenLayout.kt
@@ -32,7 +32,7 @@ fun ListScreenLayout(
     onDeleteItem: () -> Unit,
     onDeleteMultipleItems: () -> Unit,
     onFabClick: () -> Unit,
-    topSlot: (@Composable LazyItemScope.() -> Unit)? = null,
+    topSlot: (@Composable LazyItemScope.(Int) -> Unit)? = null,
     lazyColumnContent: LazyListScope.() -> Unit,
     bottomSlot: (@Composable BoxScope.() -> Unit)? = null
 ) {
